Add entities, adjust models, update repo & main
Register Libro and Autor as JPA entities and update mappings; fix ManyToMany mappedBy names and add table names. Change download_count type from Double to Integer in DatosLibro and Libro and update constructor/getter/setter accordingly. Refactor Principal to initialize Scanner/consumers, rename methods/vars, and build correct Gutendex search URL; save retrieved Libro via repository. Clean up LibroRepository: remove/rename obsolete methods, add findByIdioma, and comment out legacy queries. Update application.properties to use the literalura database name and wire LibroRepository into the application with @Autowired.

+++ src/main/java/com/alura/literalura/model/DatosLibro.java
@@ -1,14 +1,10 @@
 package com.alura.literalura.model;
-
 import com.fasterxml.jackson.annotation.JsonAlias;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import jakarta.persistence.*;
-
-import java.util.List;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public record DatosLibro(
         @JsonAlias("title") String titulo,
-        @JsonAlias("download_count") Double numeroDescargas,
+        @JsonAlias("download_count") Integer numeroDescargas,
         @JsonAlias("languages") Idioma idioma) {
 }
